pyproject.toml
[project]
name = "svec-experiment"
version = "0.1.0"
description = "Self-Verified Efficient Consistency (SVEC) prompt tuning experiment"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.14.0",
    "google-generativeai>=0.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
